refactor(link): Extract LinkConfigurationLoader
