+++ paymentsheet/src/main/java/com/stripe/android/link/injection/LinkControllerModule.kt
@@ -7,6 +7,8 @@ import com.stripe.android.common.di.MobileSessionIdModule
 import com.stripe.android.core.injection.CoreCommonModule
 import com.stripe.android.core.injection.CoroutineContextModule
 import com.stripe.android.googlepaylauncher.injection.GooglePayLauncherModule
+import com.stripe.android.link.DefaultLinkConfigurationLoader
+import com.stripe.android.link.LinkConfigurationLoader
 import com.stripe.android.paymentelement.confirmation.injection.ExtendedPaymentElementConfirmationModule
 import com.stripe.android.payments.core.injection.PRODUCT_USAGE
 import com.stripe.android.payments.core.injection.StripeRepositoryModule
@@ -15,6 +17,7 @@ import com.stripe.android.paymentsheet.injection.LinkHoldbackExposureModule
 import com.stripe.android.paymentsheet.injection.PaymentSheetCommonModule
 import com.stripe.android.ui.core.di.CardScanModule
 import com.stripe.android.ui.core.forms.resources.injection.ResourceRepositoryModule
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import javax.inject.Named
@@ -38,18 +41,24 @@ import javax.inject.Singleton
         LinkControllerComponent::class
     ]
 )
-internal object LinkControllerModule {
-    @Provides
+internal interface LinkControllerModule {
+    @Binds
     @Singleton
-    fun provideAppContext(application: Application): Context = application.applicationContext
+    fun bindLinkConfigurationLoader(impl: DefaultLinkConfigurationLoader): LinkConfigurationLoader
 
-    // TODO
-    @Provides
-    @Singleton
-    fun provideEventReporterMode(): EventReporter.Mode = EventReporter.Mode.Custom
+    companion object {
+        @Provides
+        @Singleton
+        fun provideAppContext(application: Application): Context = application.applicationContext
 
-    @Provides
-    @Singleton
-    @Named(PRODUCT_USAGE)
-    fun provideProductUsageTokens() = setOf("LinkPaymentMethodLauncher")
+        // TODO
+        @Provides
+        @Singleton
+        fun provideEventReporterMode(): EventReporter.Mode = EventReporter.Mode.Custom
+
+        @Provides
+        @Singleton
+        @Named(PRODUCT_USAGE)
+        fun provideProductUsageTokens() = setOf("LinkPaymentMethodLauncher")
+    }
 }
